Variational Smoother: QUAD4 and TRI3 tests working
Previously, attempting to run the tests resulted in a segfault. The issue was
traced to the fact that the smoother's (const) _dim attribute was initialized
based on the given mesh before the tester had finished setting up said mesh.
Thus, the _dim attribute held the wrong value, resulting in a segfault when
trying to access a vector element that did not exist.

The issue has been (for now) fixed by reinitializing _dim at the beginning of
the smooth method to ensure that the correct dimension is used.

Some sliding boundary nodes in the TRI3 tests were not honoring their
constraints to keep them on their respective boundaries. The issue was that
neighbor nodes on adjacent boundaries were being used to formulate the
constraints.

The fix was to add a check to ensure that only neighbor nodes that
share a boundary id with the constrained node are used to formulate the
constraint.

Ref. #4082