Don't renumber_nodes_and_elements to change order
This might have been necessary with the original version of these
methods, which probably predate e.g. the introduction of `max_elem_id()`
to query on meshes where that might differ from `n_elem()`, but it
shouldn't be necessary now, and it can *break* the cases where we
construct a range prior to doing renumbering and thereby invalidate the
range (e.g. moving valid elements to before the range beginning)