Move n_quadrature_points() to FEAbstract, fix it
We weren't doing anything to track quadrature point counts in cases
where we supplied a vector of points rather than relied on the qrule,
and this was causing n_quadrature_points() to return erroneous values.

This was causing errors in JumpErrorEstimator subclasses, including a
valgrind failure in one of our slit mesh unit tests.