Consider minimum node separation distance among boundary NodeElems when computing h_min
We previously used the first non-zero Elem::hmin() value encountered
in the entire Mesh to set a "preliminary" h_min value, but this
approach had a couple drawbacks: (i) it could set an h_min value from
a part of the mesh not relevant to the part actually being stitched,
and (ii) it was not sufficient in the case where the entire mesh was
comprised of NodeElems

The approach that we use now of computing the minimum node separation
distance amongst all NodeElems on the boundary should give a much more
consistent value for h_min. In the case where the boundary being
stitched consists of a single NodeElem (hence there is no minimum node
separation or valid h_min value) we will still fall back on using an
absolute tolerance check as we did previously.

* Clarify print statement. If we have not yet determined an h_min
  value, it could be because there is only a single NodeElem in the
  boundary nodeset being stitched.
* Add comment regarding user tolerance in the use_binary_search code path
* Drop preliminary h_min computation when looping over boundary node list
  Previously this was done to handle the case where a mesh contained only
  NodeElems, but we now handle that in a more consistent way later in the
  same function.