Move non-nodes-and-elements copying
We want to copy policies in our copy constructors, but *only* there, not
in the method that specifically says it will not copy most "high level"
data.

This fixes a bug in the MOOSE reactor module for me.

Unfortunately, this behavior is a 13 year old bug; hopefully other code
hasn't come to depend on it.