Iterate over GhostingFunctor* in vectors, not sets
It's entirely fair for GhostingFunctor subclasses to want to do parallel
operations in e.g. mesh_reinit(), but we can't do that if they're not
executed in the same order on all processors, and a set of pointers
is definitely not guaranteed to be in the same order on all processors.

I think I may have actually fixed this bug before anyone got bitten by
it (while investigating other issues, though).

This is technically an API change, but hopefully everybody uses "auto"
for any ugly iterators that they have to declare, in which case this
won't be an API change that breaks anyone's compile.