Improved the VariationalMeshSmoother tester
Previously, some nodes lay on a plane of symmetry and were not distorted
by the DistortHypercube class. A check has been added for this and if a
node will not be distorted due to symmetry, a 5% distortion is forced.
Because of this, we no longer have to skip checks in distortion_is on
nodes with a coordinate value of 0.5.