Utility::contains(), Utility::ends_with()
You'd think these would be easy enough to implement ourselves, but we
had a bug in our idiom for ends_with(), so let's just consolidate
everything.

I'll give these names to match C++20/23 standard names, both for an
easier transition in the future and to reassure myself that I'm not the
only one who was bad enough with string utilities to benefit from these
helper functions.