Refactored VariationalMeshSmoother
The VariationalMeshSmoother now stores as attributes: EquationSystems object,
VariationalSmootherSystem object, VariationalSmootherConstraint object, and
mesh copy object. A setup method was added to the class to facilitate the
creation and storage of these objects outside the smooth method. The motivation
for this is that, by having these objects live outside the solve method, other
code is able to query the VariationalSmootherSystem for it’s _mesh_info
attribute to asses the pre and post-smoothed quality of the mesh. This is
immediately useful in the unit tests where we will want to verify that our
distortion has tangled the mesh.