Add PerfLog to Netgen interface
This didn't end up giving me anything *actionable*, since our issues
with Valgrind+Netgen turn out to be related to some heinous interaction
they have while Netgen spins up threads (3 threads, and they don't
expose an API to control that).

But this is still an expensive function with a couple big parts that are
worth logging.