Default to sync c++ streams
This may be slower but Roy and I think it makes more sense for the
default to be safety rather than speed. Safety here means that we
guarantee no data races to the stream. You can still have interleaved
bytes (for which we are planning a new thread safe ostream proxy)
but at least each byte write will be atomic