Skip remote children in bracketing_nodes search
This fixes a downstream failure for me doing adaptivity with Quad8
elements, where the nasty search we do to find bracketing nodes of
incomplete-order elmeents can hit remote children if they're ghosted.

I'm not completely confident this fixes the problem without just opening
us up to even harder-to-trigger problems, but it's at least an
improvement.