Avoid trivial memory leak in unit test driver
If we add only a subset of tests to the runner (via --re or --deny_re
options), we're careful to move the other tests to a "rejects" suite so
they'll get deleted there, but the "supertest", the suite holding all
the other tests, wasn't getting deleted in those cases.

This (on top of my earlier fixes, and using the MOOSE suppressions file
for third-party library issues) gets selective valgrind runs of our unit
tests clean for me.