Drop deprecated DifferentiableQoI::init_qoi()
This virtual keyword on this function was deprecated in 6fbd01f6
(2022), so it has been deprecated since the 1.8.x release series. The
non-deprecated, non-virtual version of the function was made into a
no-op. Therefore, there may be code which still calls this function,
but it would just be doing nothing, so let's try removing it to see
whether everything still compiles.