Drop deprecated libmesh_make_unique macro
This was originally a workaround for compilers that did not support
std::make_unique, but it should not be getting used by anything now.

Using this macro was deprecated in 55813c4b (2022), which appeared in
the 1.8.x release series, so now is a good time to get rid of it.