Drop private: declaration wrapped in LIBMESH_ENABLE_DEPRECATED guards
This private section has been deprecated since 68bd377a (2020, 1.7.x
series), but it's not clear to me that this was doing anything even
when LIBEMSH_ENABLE_DEPRECATED was defined, since there is a private:
declaration a few lines higher up in the class. So it should be
totally safe to get rid of it.