Call ExodusII_IO_Helper::get_libmesh_node_id() from ExodusII_IO::read()
The helper function effectively does nothing in the case where
set_unique_ids_from_maps == true, since in that case the libmesh id is
just a (zero-based) version of the input. We then determine the
unique_id after the new Node has already been added, since that is
when it is most convenient to set the unique_id value.