Drop long deprecated System::read_legacy_data()
This has been deprecated since before 3f8ade9cc (Dec 2011). The
history prior to that commit is a bit difficult to interpret since it
coincides with the time we switched from svn to git, but that function
also appears to have been deprecated since before c3ef1180 (Oct 2009)
when the "systems" directory was initially created.