Convert unit test exo files from netCDF-4 to 64-bit offset
To do the conversion, first ncdump the file to text:

    $LIBMESH_DIR/bin/ncdump foo.e > foo.txt

Edit the resulting text file so that it does not contain any features
not supported by older Exodus files. For example:

    :file_size = 1LL ;              --> :file_size = 1 ;
    int64 ns_prop1(num_node_sets) ; --> int ns_prop1(num_node_sets) ;

Then ncgen with the -k '64-bit offset' option:

    $LIBMESH_DIR/bin/ncgen foo.txt -k '64-bit offset' -o foo.e

The ncdump and ncgen commands are located in the bin/ directory of
your libmesh installation.